Removed Custom Instana propagator for baggage propagation since multitextmappropagator will suffice the same
