+++ src/Propagation/Instana/tests/Unit/InstanaPropagatorTest.php
@@ -13,14 +13,15 @@ use OpenTelemetry\API\Trace\TraceFlags;
 use OpenTelemetry\Context\Context;
 
 use OpenTelemetry\Context\ContextInterface;
-use OpenTelemetry\Contrib\Propagation\Instana\InstanaContextPropagator as InstanaContextPropagator;
-use OpenTelemetry\Contrib\Propagation\Instana\InstanaPropagator as InstanaPropagator;
+use OpenTelemetry\Context\Propagation\MultiTextMapPropagator;
+use OpenTelemetry\Contrib\Propagation\Instana\InstanaPropagator;
+use OpenTelemetry\API\Baggage\Propagation\BaggagePropagator;
 
 use OpenTelemetry\SDK\Trace\Span;
 use Override;
 use PHPUnit\Framework\TestCase;
 
-final class InstanaContextPropagatorTest extends TestCase
+final class InstanaPropagatorTest extends TestCase
 {
     private const X_INSTANA_T = 'ff000000000000000000000000000041';
     private const X_INSTANA_S = 'ff00000000000041';
@@ -31,13 +32,13 @@ final class InstanaContextPropagatorTest extends TestCase
     private $SPAN_ID;
     private $SAMPLED;
 
-    private InstanaContextPropagator $InstanaContextPropagator;
+    private InstanaPropagator $InstanaPropagator;
     
     #[Override]
     protected function setUp(): void
     {
-        $this->InstanaContextPropagator = InstanaContextPropagator::getInstance();
-        $instanaMultiFields = $this->InstanaContextPropagator->fields();
+        $this->InstanaPropagator = InstanaPropagator::getInstance();
+        $instanaMultiFields = $this->InstanaPropagator->fields();
         $this->TRACE_ID = $instanaMultiFields[0];
         $this->SPAN_ID = $instanaMultiFields[1];
         $this->SAMPLED = $instanaMultiFields[2];
@@ -47,14 +48,14 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $this->assertSame(
             ['X-INSTANA-T', 'X-INSTANA-S',  'X-INSTANA-L'],
-            $this->InstanaContextPropagator->fields()
+            $this->InstanaPropagator->fields()
         );
     }
 
     public function test_inject_empty(): void
     {
         $carrier = [];
-        $this->InstanaContextPropagator->inject($carrier);
+        $this->InstanaPropagator->inject($carrier);
         $this->assertEmpty($carrier);
     }
 
@@ -62,7 +63,7 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $carrier = [];
         $this
-            ->InstanaContextPropagator
+            ->InstanaPropagator
             ->inject(
                 $carrier,
                 null,
@@ -82,7 +83,7 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $carrier = [];
         $this
-            ->InstanaContextPropagator
+            ->InstanaPropagator
             ->inject(
                 $carrier,
                 null,
@@ -106,7 +107,7 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $carrier = [];
         $this
-            ->InstanaContextPropagator
+            ->InstanaPropagator
             ->inject(
                 $carrier,
                 null,
@@ -130,7 +131,7 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $carrier = [];
         $this
-            ->InstanaContextPropagator
+            ->InstanaPropagator
             ->inject(
                 $carrier,
                 null,
@@ -160,7 +161,7 @@ final class InstanaContextPropagatorTest extends TestCase
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S, TraceFlags::SAMPLED),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -185,12 +186,12 @@ final class InstanaContextPropagatorTest extends TestCase
             $this->SAMPLED => $sampledValue,
             'baggage' => 'user_id=12345,request_id=abcde',
         ];
-        $instanaPropagator = InstanaPropagator::getInstance();
-        $context = $instanaPropagator->extract($carrier);
+        $propagator = new MultiTextMapPropagator([InstanaPropagator::getInstance(), BaggagePropagator::getInstance()]);
+        $context = $propagator->extract($carrier);
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S, TraceFlags::SAMPLED),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
 
         // Verify baggage
@@ -221,11 +222,11 @@ final class InstanaContextPropagatorTest extends TestCase
             $this->SAMPLED => $sampledValue,
             'baggage' => 'user_id=12345,request_id=abcde',
         ];
-        $context = $this->InstanaContextPropagator->extract($carrier);
+        $context = $this->InstanaPropagator->extract($carrier);
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S, TraceFlags::SAMPLED),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
 
         // Verify baggage is not propagated
@@ -239,7 +240,9 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $carrier = [];
 
-        InstanaPropagator::getInstance()->inject(
+        $propagator = new MultiTextMapPropagator([InstanaPropagator::getInstance(), BaggagePropagator::getInstance()]);
+
+        $propagator->inject(
             $carrier,
             null,
             Context::getRoot()->withContextValue(
@@ -269,7 +272,7 @@ final class InstanaContextPropagatorTest extends TestCase
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S, TraceFlags::SAMPLED),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -296,7 +299,7 @@ final class InstanaContextPropagatorTest extends TestCase
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -323,7 +326,7 @@ final class InstanaContextPropagatorTest extends TestCase
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S, TraceFlags::DEFAULT),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -348,7 +351,7 @@ final class InstanaContextPropagatorTest extends TestCase
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent(self::X_INSTANA_T, self::X_INSTANA_S),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -372,7 +375,7 @@ final class InstanaContextPropagatorTest extends TestCase
                 '0000000000000000',
                 TraceFlags::SAMPLED
             ),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -388,7 +391,7 @@ final class InstanaContextPropagatorTest extends TestCase
                 '0000000000000000',
                 TraceFlags::DEFAULT
             ),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -399,9 +402,9 @@ final class InstanaContextPropagatorTest extends TestCase
             $this->SPAN_ID => self::X_INSTANA_S,
             $this->SAMPLED => self::IS_SAMPLED,
         ];
-        $context = $this->InstanaContextPropagator->extract($extractCarrier);
+        $context = $this->InstanaPropagator->extract($extractCarrier);
         $injectCarrier = [];
-        $this->InstanaContextPropagator->inject($injectCarrier, null, $context);
+        $this->InstanaPropagator->inject($injectCarrier, null, $context);
         $this->assertSame($injectCarrier, $extractCarrier);
     }
 
@@ -426,7 +429,7 @@ final class InstanaContextPropagatorTest extends TestCase
 
         $this->assertEquals(
             SpanContext::createFromRemoteParent('00000000000000004aaba1a52cf8ee09', '0007b5a2e4d86bd1', TraceFlags::SAMPLED),
-            $this->getSpanContext($this->InstanaContextPropagator->extract($carrier))
+            $this->getSpanContext($this->InstanaPropagator->extract($carrier))
         );
     }
 
@@ -489,7 +492,7 @@ final class InstanaContextPropagatorTest extends TestCase
     {
         $this->assertSame(
             Context::getCurrent(),
-            $this->InstanaContextPropagator->extract($carrier),
+            $this->InstanaPropagator->extract($carrier),
         );
     }
